fix bug found in pangoterm, make vt parsing sligthly less janky
after sigwinch this sets a flag asynchronously, then it used to read
all the input for the current frame, and finally it would check the flag
and handle the resize event by setting the cursor position to 9999;9999
and asking for the new cursos position.  for some reason in pangoterm
this would semi reliably send a focus off event before the cursor
position.  this can only happen if the focus event happened after
reading the input but before querying the terminal, which seems crazy
unlikely, but pangoterm was hitting it almost every time

so the problem really is that get_term_size was doing its own reading
and parsing.  instead, let's query the position before reading and use
the normal parsing after the input is read